Get signature gene identifiers

Obtain gene signatures implemented in hacksig as a named list of gene symbols.

Usage

get_sig_genes(keywords = "all")

Arguments

keywords

A character vector indicating keywords for a group of signatures. The default ("all") will cause the function to check for all the signatures implemented in hacksig.

Value

A named list of gene signatures.

See also

Examples

get_sig_genes()
#> $ayers2017_immexp
#>  [1] "CCL5"    "CD2"     "CD3D"    "CD3E"    "CIITA"   "CXCL10"  "CXCL13" 
#>  [8] "CXCR6"   "GZMB"    "GZMK"    "HLA-DRA" "HLA-E"   "IDO1"    "IL2RG"  
#> [15] "LAG3"    "NKG7"    "STAT1"   "TAGAP"  
#> 
#> $bai2019_immune
#>  [1] "BATF"    "CCL11"   "CCR4"    "CCR7"    "CD27"    "CD79B"   "CMA1"   
#>  [8] "CNR2"    "CTLA4"   "CTSG"    "GZMM"    "IL16"    "IL19"    "MASP1"  
#> [15] "PGLYRP4" "SAA1"    "TNFAIP3" "TREML1" 
#> 
#> $cinsarc
#>  [1] "ANLN"     "ASPM"     "AURKA"    "AURKB"    "BIRC5"    "BORA"    
#>  [7] "BUB1"     "BUB1B"    "CCNA2"    "CCNB1"    "CCNB2"    "CDC20"   
#> [13] "CDC45"    "CDC6"     "CDC7"     "CDCA2"    "CDCA3"    "CDCA8"   
#> [19] "CDK1"     "CENPA"    "CENPE"    "CENPL"    "CEP55"    "CHEK1"   
#> [25] "CKAP5"    "CKS2"     "ECT2"     "ESPL1"    "FANCI"    "FBXO5"   
#> [31] "FOXM1"    "H2AX"     "HP1BP3"   "KIF11"    "KIF14"    "KIF15"   
#> [37] "KIF18A"   "KIF20A"   "KIF23"    "KIF2C"    "KIF4A"    "MAD2L1"  
#> [43] "MCM2"     "MCM7"     "MELK"     "NCAPH"    "NDE1"     "NEK2"    
#> [49] "NUF2"     "OIP5"     "PBK"      "PLK4"     "PRC1"     "PTTG1"   
#> [55] "RAD51AP1" "RNASEH2A" "RRM2"     "SGO2"     "SMC2"     "SPAG5"   
#> [61] "SPC25"    "TOP2A"    "TPX2"     "TRIP13"   "TTK"      "UBE2C"   
#> [67] "ZWINT"   
#> 
#> $dececco2014_int172
#>   [1] "ABI2"       "ACSL3"      "AKNA"       "ALS2"       "AMPD3"     
#>   [6] "AP3D1"      "ARHGAP30"   "ARHGEF40"   "ARMCX1"     "ARMCX4"    
#>  [11] "B4GAT1"     "BBS9"       "BCL2L1"     "BEX3"       "BOD1"      
#>  [16] "C11orf49"   "CACTIN"     "CADM1"      "CAMLG"      "CARMIL2"   
#>  [21] "CARS2"      "CCDC43"     "CD101"      "CD200R1"    "CD6"       
#>  [26] "CD84"       "CENPBD1"    "CFDP1"      "CHSY3"      "CKAP4"     
#>  [31] "COPS8"      "COX6B2"     "CPNE9"      "CXCR3"      "DAPK3"     
#>  [36] "DMTN"       "DOC2B"      "DYRK4"      "EFNB3"      "EHMT1"     
#>  [41] "EXTL2"      "FAM20C"     "FAM53C"     "FASTKD2"    "FGD2"      
#>  [46] "FIBCD1"     "FLI1"       "FLOT2"      "FLT3LG"     "FOXP3"     
#>  [51] "FSTL1"      "FXYD2"      "GBP4"       "GCH1"       "GNA13"     
#>  [56] "GNG8"       "GPRASP2"    "GRK2"       "GVINP1"     "GZF1"      
#>  [61] "HS3ST4"     "HTRA3"      "IFI16"      "ILKAP"      "IQGAP1"    
#>  [66] "ITGB5"      "JAM3"       "KCNE5"      "KEL"        "KIR2DL4"   
#>  [71] "KIR2DS2"    "KITLG"      "KLRC2"      "LAMB1"      "LAMC1"     
#>  [76] "LDLRAD3"    "LILRA6"     "LIMK1"      "LMAN2L"     "LOC153684" 
#>  [81] "LSP1P5"     "LYZ"        "METTL7B"    "MRFAP1"     "MUC16"     
#>  [86] "MYH10"      "NADK"       "NAP1L5"     "NCL"        "NLN"       
#>  [91] "NLRC5"      "NPIPB5"     "NPR1"       "NUDT9"      "PALM2AKAP2"
#>  [96] "PCGF2"      "PDE6D"      "PDHX"       "PGBD1"      "PHC1"      
#> [101] "PHETA2"     "PLEKHM3"    "PMM1"       "POFUT1"     "POGLUT2"   
#> [106] "PRF1"       "PRKD1"      "PSD4"       "PTK2B"      "PTX3"      
#> [111] "PXDN"       "RAP2B"      "RASSF5"     "RASSF8"     "RCN1"      
#> [116] "RESF1"      "RETREG2"    "RHOF"       "RNF24"      "RSPH6A"    
#> [121] "SCFD2"      "SH3RF3"     "SLC15A1"    "SLC22A17"   "SLC37A3"   
#> [126] "SLFN5"      "SMARCA1"    "SNCG"       "SNX20"      "SNX21"     
#> [131] "SNX29P2"    "SORBS1"     "SPATS2L"    "SPI1"       "SPOCK1"    
#> [136] "SPRY1"      "SRPRA"      "ST6GALNAC4" "ST7-AS1"    "ST8SIA2"   
#> [141] "STAU1"      "STBD1"      "STK25"      "SV2A"       "SYNGR3"    
#> [146] "SYNJ1"      "TBC1D20"    "TCEAL4"     "TCEAL9"     "THUMPD2"   
#> [151] "TIGD7"      "TLR6"       "TLR8"       "TMC8"       "TMCO1-AS1" 
#> [156] "TMEM128"    "TMEM259"    "TNFAIP8"    "TNN"        "TPST1"     
#> [161] "TRAM1L1"    "TREML2"     "UBAC2"      "UBXN6"      "USP13"     
#> [166] "USP40"      "VASP"       "VAT1"       "VPS26C"     "XKR6"      
#> [171] "ZC3H12B"    "ZDHHC9"    
#> 
#> $eschrich2009_rsi
#>  [1] "ABL1"  "AR"    "HDAC1" "IRF1"  "JUN"   "PAK2"  "PRKCB" "RELA"  "STAT1"
#> [10] "SUMO1"
#> 
#> $estimate_immune
#>   [1] "ADAM8"    "ADCY7"    "ALOX5AP"  "AOAH"     "ARHGAP15" "ARHGDIB" 
#>   [7] "ARHGEF6"  "BCL2A1"   "BIN2"     "CASP1"    "CCDC69"   "CCL5"    
#>  [13] "CCR7"     "CD2"      "CD247"    "CD27"     "CD300A"   "CD302"   
#>  [19] "CD37"     "CD3D"     "CD48"     "CD52"     "CD53"     "CD69"    
#>  [25] "CD74"     "CLEC2B"   "CLEC4A"   "CORO1A"   "CST7"     "CSTA"    
#>  [31] "CTSS"     "CYBB"     "DOCK2"    "EMP3"     "EVI2B"    "FCER1G"  
#>  [37] "FGL2"     "FGR"      "FLI1"     "FYB1"     "GBP1"     "GBP2"    
#>  [43] "GIMAP4"   "GIMAP6"   "GLRX"     "GMFG"     "GMIP"     "GNLY"    
#>  [49] "GPR65"    "GZMB"     "GZMH"     "GZMK"     "HCK"      "HCLS1"   
#>  [55] "HLA-B"    "HLA-DMA"  "HLA-DPA1" "HLA-DPB1" "HLA-DRA"  "HLA-E"   
#>  [61] "HLA-F"    "HLA-G"    "IFI30"    "IKZF1"    "IL10RA"   "IL18RAP" 
#>  [67] "IL2RB"    "IL2RG"    "IL32"     "IL4R"     "IL7R"     "IRF8"    
#>  [73] "ITGA4"    "ITGAL"    "ITGB2"    "ITK"      "KLRB1"    "KLRK1"   
#>  [79] "LAIR1"    "LAPTM5"   "LCK"      "LCP2"     "LGALS9"   "LILRB1"  
#>  [85] "LILRB2"   "LPXN"     "LSP1"     "LST1"     "LTB"      "LY96"    
#>  [91] "LYZ"      "MAFB"     "MFSD1"    "MICAL1"   "MNDA"     "MSN"     
#>  [97] "MYO1F"    "NCF2"     "NCF4"     "NCKAP1L"  "NFKBIA"   "NKG7"    
#> [103] "P2RY14"   "PLEK"     "PRF1"     "PSTPIP1"  "PTGER2"   "PTGER4"  
#> [109] "PTPRC"    "PTPRCAP"  "PTPRE"    "PVRIG"    "RAB27A"   "RABGAP1L"
#> [115] "RAC2"     "RASSF2"   "RGS1"     "RHOG"     "RHOH"     "RNASE6"  
#> [121] "S100A8"   "S100A9"   "SAMHD1"   "SELL"     "SELPLG"   "SH2B3"   
#> [127] "SLA"      "SRGN"     "TAP1"     "TCIRG1"   "TGFB1"    "TNFAIP3" 
#> [133] "TNFRSF1B" "TPP1"     "TPST2"    "TRAF3IP3" "TYROBP"   "VAV1"    
#> [139] "VNN2"     "WIPF1"    "ZAP70"   
#> 
#> $estimate_stromal
#>   [1] "ABCA6"    "ACTG2"    "ADAM12"   "ADAMTS5"  "ADGRA2"   "AIF1"    
#>   [7] "AOC3"     "APBB1IP"  "ARHGAP28" "ASPN"     "ATP8B4"   "BGN"     
#>  [13] "BTK"      "C1QA"     "C1QB"     "C3AR1"    "CCN4"     "CCR1"    
#>  [19] "CD14"     "CD163"    "CD200"    "CD248"    "CD33"     "CD86"    
#>  [25] "CD93"     "CDH5"     "CH25H"    "CILP"     "CLEC7A"   "COL10A1" 
#>  [31] "COL14A1"  "COL15A1"  "COL1A2"   "COL3A1"   "COL5A3"   "COL6A3"  
#>  [37] "COL8A2"   "COMP"     "COX7A1"   "CSF1R"    "CXCL12"   "CXCL14"  
#>  [43] "CXCL9"    "DCN"      "DDR2"     "DIO2"     "ECM2"     "EDIL3"   
#>  [49] "EDNRA"    "EGFL6"    "EMCN"     "ENPEP"    "ENPP2"    "ERG"     
#>  [55] "F13A1"    "FAP"      "FASLG"    "FBLN2"    "FCGR2A"   "FCGR2B"  
#>  [61] "FMO1"     "FOXF1"    "FPR1"     "FRZB"     "GIMAP5"   "GREM1"   
#>  [67] "HDC"      "HEPH"     "HGF"      "IGF1"     "IGSF6"    "IL18R1"  
#>  [73] "IL1B"     "ISLR"     "ITGAM"    "ITGBL1"   "ITIH3"    "ITIH5"   
#>  [79] "ITM2A"    "KCNJ8"    "KDR"      "LDB2"     "LMOD1"    "LRRC15"  
#>  [85] "LRRC32"   "LUM"      "LY86"     "MAF"      "MFAP5"    "MMP3"    
#>  [91] "MS4A4A"   "MS4A6A"   "MSR1"     "MXRA5"    "NME8"     "NOX4"    
#>  [97] "OLFML1"   "OLFML2B"  "PAPPA"    "PCDH12"   "PDE2A"    "PDGFRB"  
#> [103] "PIK3R5"   "PLPPR4"   "PLXDC1"   "PLXNC1"   "PRKG1"    "PTGER3"  
#> [109] "PTGIS"    "RAMP3"    "RARRES2"  "RASGRP3"  "RGS4"     "RSAD2"   
#> [115] "RUNX1T1"  "SAMSN1"   "SCUBE2"   "SERPING1" "SFRP4"    "SGCD"    
#> [121] "SH2D1A"   "SIGLEC1"  "SP140"    "SPON1"    "SPON2"    "SULF1"   
#> [127] "TCF21"    "TENM4"    "TFEC"     "THBS2"    "TLR2"     "TLR7"    
#> [133] "TNFSF4"   "TNN"      "TRAT1"    "VCAM1"    "VSIG4"    "WNT2"    
#> [139] "ZEB2"     "ZFPM2"    "ZNF423"  
#> 
#> $eustace2013_hypoxia
#>  [1] "ALDOA"   "ANGPTL4" "ANLN"    "BNC1"    "CA9"     "CDKN3"   "COL4A6" 
#>  [8] "DCBLD1"  "ENO1"    "FAM83B"  "FOSL1"   "GNAI1"   "HILPDA"  "KCTD11" 
#> [15] "KRT17"   "LDHA"    "MRGBP"   "MRPS17"  "P4HA1"   "PGAM1"   "PGK1"   
#> [22] "SDC1"    "SLC16A1" "SLC2A1"  "TPI1"    "VEGFA"  
#> 
#> $fan2021_ferroptosis
#>  [1] "ASNS"      "ATG5"      "AURKA"     "BAP1"      "BNIP3"     "CISD2"    
#>  [7] "DRD4"      "FBXW7"     "FTH1"      "LINC00336" "MAP1LC3A"  "MAP3K5"   
#> [13] "PRDX6"     "SLC2A3"    "SLC7A5"    "SOCS1"     "ZFP69B"   
#> 
#> $fang2021_irgs
#>  [1] "ADM"       "BTC"       "CTLA4"     "DEFB1"     "DKK1"      "EDNRB"    
#>  [7] "FAM3D"     "GNRH1"     "STC2"      "TNFRSF12A"
#> 
#> $han2021_ferroptosis
#> [1] "FTH1"  "PHKG2" "TFRC" 
#> 
#> $he2021_ferroptosis_a
#> [1] "CDKN2A"  "EIF2S1"  "FTH1"    "KEAP1"   "MAP3K5"  "SELENOS" "SLC2A3" 
#> 
#> $he2021_ferroptosis_b
#>  [1] "ACSL3"    "ATG5"     "DDIT4"    "MAP1LC3A" "MAP3K5"   "OTUB1"   
#>  [7] "PRDX6"    "PRKAA2"   "SLC7A5"   "SOCS1"   
#> 
#> $hu2021_derbp
#> [1] "AZGP1"   "EZH2"    "IGF2BP2" "MKRN3"   "MRPL47"  "NCBP2"  
#> 
#> $huang2022_ferroptosis
#> [1] "ATG5"   "CDKN2A" "MAP3K5" "OTUB1"  "SLC2A3" "SOCS1"  "TRIB3" 
#> 
#> $ips_cp
#>  [1] "CD27"     "CD274"    "CTLA4"    "HAVCR2"   "ICOS"     "IDO1"    
#>  [7] "LAG3"     "PDCD1"    "PDCD1LG2" "TIGIT"   
#> 
#> $ips_ec
#>   [1] "ACAP1"     "ADRM1"     "AHSA1"     "AIM2"      "APOL3"     "ARHGAP10" 
#>   [7] "ATM"       "ATP10D"    "BIRC3"     "BRIP1"     "C1GALT1C1" "C3AR1"    
#>  [13] "CASP3"     "CASQ1"     "CAVIN2"    "CCL20"     "CCL4"      "CCL5"     
#>  [19] "CCNB1"     "CCR5"      "CCR7"      "CCT6B"     "CD160"     "CD300E"   
#>  [25] "CD37"      "CD3D"      "CD3E"      "CD3G"      "CD55"      "CD69"     
#>  [31] "CD8A"      "CETN3"     "CFLAR"     "CMKLR1"    "CSE1L"     "DAPP1"    
#>  [37] "DARS1"     "DOCK9"     "DUSP2"     "ESCO2"     "ETS1"      "EXO1"     
#>  [43] "EXOC6"     "EXOSC9"    "EZH2"      "FCRL6"     "FLT3LG"    "GDE1"     
#>  [49] "GEMIN6"    "GNLY"      "GPT2"      "GZMA"      "GZMH"      "GZMK"     
#>  [55] "GZMM"      "HAPLN3"    "HLA-DMB"   "HLA-DPA1"  "HLA-DPB1"  "IARS1"    
#>  [61] "IFI16"     "IL2RB"     "IL34"      "KIF11"     "KNTC1"     "LCK"      
#>  [67] "LIME1"     "LTK"       "MPZL1"     "NCOA4"     "NEFL"      "NFKBIA"   
#>  [73] "NKG7"      "NUF2"      "PDGFRL"    "PIK3IP1"   "PRC1"      "PSAT1"    
#>  [79] "PTGS1"     "PTRH2"     "REPS1"     "RGS1"      "RTKN2"     "SAMSN1"   
#>  [85] "SCG2"      "SELL"      "SETD7"     "SIGLEC14"  "SIGLEC6"   "SIK1"     
#>  [91] "TAL1"      "TFEC"      "TIMM13"    "TIPIN"     "TPK1"      "TRAT1"    
#>  [97] "TRIB2"     "UQCRB"     "USP9Y"     "WIPF1"     "ZAP70"     "ZCRB1"    
#> 
#> $ips_mhc
#>  [1] "B2M"      "HLA-A"    "HLA-B"    "HLA-C"    "HLA-DPA1" "HLA-DPB1"
#>  [7] "HLA-E"    "HLA-F"    "TAP1"     "TAP2"    
#> 
#> $ips_sc
#>  [1] "CCL3L1"  "CCR2"    "CD14"    "CD2"     "CD72"    "CD86"    "CLEC5A" 
#>  [8] "CXCR4"   "FCGR2A"  "FCGR2B"  "FCGR3A"  "FERMT3"  "FOXP3"   "GPSM3"  
#> [15] "IL18BP"  "IL4R"    "ITGA4"   "ITGAL"   "ITGAM"   "L1CAM"   "LIPA"   
#> [22] "LRP1"    "LRRC42"  "MARCO"   "MMP12"   "MNDA"    "MRC1"    "MS4A6A" 
#> [29] "PARVG"   "PELO"    "PLEK"    "PRSS23"  "PSAP"    "PTGER2"  "PTGES2" 
#> [36] "PTGIR"   "S100A8"  "S100A9"  "ST8SIA4" "STAB1"  
#> 
#> $li2021_ferroptosis_a
#>  [1] "ACO1"     "AKR1C3"   "ALOX15"   "ATG5"     "BID"      "GLS2"    
#>  [7] "GOT1"     "MAP1LC3A" "MAP3K5"   "SCO2"    
#> 
#> $li2021_ferroptosis_b
#>  [1] "ASNS"   "AURKA"  "CAV1"   "CDKN2A" "CISD2"  "DDIT4"  "EGFR"   "G6PD"  
#>  [9] "PRKAA2" "SLC7A5" "SOCS1"  "TRIB3" 
#> 
#> $li2021_ferroptosis_c
#> [1] "AKR1C3"  "AURKA"   "CAV1"    "SLC7A11" "TRIB3"  
#> 
#> $li2021_ferroptosis_d
#> [1] "FLT3"  "IL6"   "KEAP1" "NQO1"  "SOCS1" "TRIB3"
#> 
#> $li2021_irgs
#> [1] "BMP7"  "PLAUR" "PRKCQ" "PSMD2"
#> 
#> $liu2020_immune
#> [1] "DKK1"      "HBEGF"     "INHBA"     "PIK3R3"    "RNASE7"    "TNFRSF12A"
#> 
#> $liu2021_mgs
#> [1] "CCL13" "CCR7"  "CD276" "IL1B"  "LYVE1" "VEGFC"
#> 
#> $lohavanichbutr2013_hpvneg
#>  [1] "CLEC3B"   "LAMC2"    "LIPI"     "MYH11"    "NREP"     "OASL"    
#>  [7] "OSMR"     "PLEKHA7"  "SERPINE1" "SLC16A1"  "THBS1"    "TPPP"    
#> [13] "ZDHHC11" 
#> 
#> $lu2020_npc
#> [1] "IGF2BP1" "IGF2BP2" "METTL3" 
#> 
#> $lu2021_ferroptosis
#> [1] "BNIP3"  "FTH1"   "SLC2A3" "TRIB3" 
#> 
#> $muro2016_ifng
#> [1] "CXCL10"  "CXCL9"   "HLA-DRA" "IDO1"    "IFNG"    "STAT1"  
#> 
#> $qiang2021_irgs
#>  [1] "BLNK"      "CCL26"     "FPR2"      "GAST"      "IL34"      "IRF9"     
#>  [7] "PLAU"      "SEMA3G"    "SLURP1"    "STC1"      "STC2"      "TNFRSF12A"
#> [13] "TNFRSF25" 
#> 
#> $rooney2015_cyt
#> [1] "GZMA" "PRF1"
#> 
#> $shan2021_ferroptosis
#>  [1] "AKR1C2"  "CD44"    "CDO1"    "CYBB"    "DUOX2"   "EGFR"    "EPAS1"  
#>  [8] "FTH1"    "MYB"     "NOX1"    "NOX3"    "PROM2"   "SLC7A11" "SQSTM1" 
#> [15] "TP63"   
#> 
#> $she2020_irgs
#>  [1] "AVPR2"    "BTC"      "CCL22"    "CCR6"     "CHGB"     "DKK1"    
#>  [7] "HBEGF"    "HRG"      "ICOS"     "IL20RA"   "INHBB"    "KLRK1"   
#> [13] "LCNL1"    "MASP1"    "OLR1"     "PDGFA"    "PTX3"     "RBP4"    
#> [19] "RFXAP"    "ROBO1"    "RORB"     "SH3BP2"   "TMSB4Y"   "TNFRSF18"
#> [25] "TNFRSF25" "TNFRSF4"  "ULBP1"   
#> 
#> $si2022_npc
#> [1] "FOXO6"    "LZTS2"    "MAPK8IP2" "SCUBE3"  
#> 
#> $tang2018_npc
#>  [1] "ANXA1"  "CBR3"   "CLASP1" "CXCL10" "DCTN1"  "FNDC3B" "GRM4"   "HDLBP" 
#>  [9] "LRIG1"  "POLR2M" "WNK1"   "WSB2"   "YBX3"  
#> 
#> $wu2020_metabolic
#>  [1] "ACAA1"    "ACAT1"    "ACOX3"    "ADA"      "ADH7"     "ADK"     
#>  [7] "AGPS"     "ASNS"     "ATIC"     "DGKQ"     "DNMT1"    "ENTPD1"  
#> [13] "GNPDA1"   "HADHB"    "HEXB"     "HPRT1"    "KYNU"     "LCLAT1"  
#> [19] "MTHFD2"   "NADSYN1"  "PAFAH1B2" "PIP4K2A"  "PLA2G2D"  "PLCB3"   
#> [25] "POLD2"    "POLE2"    "PRPS1"    "PYGL"     "RDH11"    "SMS"     
#> 
#> $xu2021_ferroptosis
#> [1] "ATG5"   "FTH1"   "MAP3K5" "OTUB1"  "PRDX6"  "SOCS1" 
#> 
#> $zhu2021_ferroptosis
#> [1] "CA9"     "NRAS"    "TNFAIP3"
#> 
#> $zou2020_npc
#> [1] "ACKR1"  "IGHG1"  "IL33"   "SLC6A8"
#> 
get_sig_genes("estimate")
#> $estimate_immune
#>   [1] "ADAM8"    "ADCY7"    "ALOX5AP"  "AOAH"     "ARHGAP15" "ARHGDIB" 
#>   [7] "ARHGEF6"  "BCL2A1"   "BIN2"     "CASP1"    "CCDC69"   "CCL5"    
#>  [13] "CCR7"     "CD2"      "CD247"    "CD27"     "CD300A"   "CD302"   
#>  [19] "CD37"     "CD3D"     "CD48"     "CD52"     "CD53"     "CD69"    
#>  [25] "CD74"     "CLEC2B"   "CLEC4A"   "CORO1A"   "CST7"     "CSTA"    
#>  [31] "CTSS"     "CYBB"     "DOCK2"    "EMP3"     "EVI2B"    "FCER1G"  
#>  [37] "FGL2"     "FGR"      "FLI1"     "FYB1"     "GBP1"     "GBP2"    
#>  [43] "GIMAP4"   "GIMAP6"   "GLRX"     "GMFG"     "GMIP"     "GNLY"    
#>  [49] "GPR65"    "GZMB"     "GZMH"     "GZMK"     "HCK"      "HCLS1"   
#>  [55] "HLA-B"    "HLA-DMA"  "HLA-DPA1" "HLA-DPB1" "HLA-DRA"  "HLA-E"   
#>  [61] "HLA-F"    "HLA-G"    "IFI30"    "IKZF1"    "IL10RA"   "IL18RAP" 
#>  [67] "IL2RB"    "IL2RG"    "IL32"     "IL4R"     "IL7R"     "IRF8"    
#>  [73] "ITGA4"    "ITGAL"    "ITGB2"    "ITK"      "KLRB1"    "KLRK1"   
#>  [79] "LAIR1"    "LAPTM5"   "LCK"      "LCP2"     "LGALS9"   "LILRB1"  
#>  [85] "LILRB2"   "LPXN"     "LSP1"     "LST1"     "LTB"      "LY96"    
#>  [91] "LYZ"      "MAFB"     "MFSD1"    "MICAL1"   "MNDA"     "MSN"     
#>  [97] "MYO1F"    "NCF2"     "NCF4"     "NCKAP1L"  "NFKBIA"   "NKG7"    
#> [103] "P2RY14"   "PLEK"     "PRF1"     "PSTPIP1"  "PTGER2"   "PTGER4"  
#> [109] "PTPRC"    "PTPRCAP"  "PTPRE"    "PVRIG"    "RAB27A"   "RABGAP1L"
#> [115] "RAC2"     "RASSF2"   "RGS1"     "RHOG"     "RHOH"     "RNASE6"  
#> [121] "S100A8"   "S100A9"   "SAMHD1"   "SELL"     "SELPLG"   "SH2B3"   
#> [127] "SLA"      "SRGN"     "TAP1"     "TCIRG1"   "TGFB1"    "TNFAIP3" 
#> [133] "TNFRSF1B" "TPP1"     "TPST2"    "TRAF3IP3" "TYROBP"   "VAV1"    
#> [139] "VNN2"     "WIPF1"    "ZAP70"   
#> 
#> $estimate_stromal
#>   [1] "ABCA6"    "ACTG2"    "ADAM12"   "ADAMTS5"  "ADGRA2"   "AIF1"    
#>   [7] "AOC3"     "APBB1IP"  "ARHGAP28" "ASPN"     "ATP8B4"   "BGN"     
#>  [13] "BTK"      "C1QA"     "C1QB"     "C3AR1"    "CCN4"     "CCR1"    
#>  [19] "CD14"     "CD163"    "CD200"    "CD248"    "CD33"     "CD86"    
#>  [25] "CD93"     "CDH5"     "CH25H"    "CILP"     "CLEC7A"   "COL10A1" 
#>  [31] "COL14A1"  "COL15A1"  "COL1A2"   "COL3A1"   "COL5A3"   "COL6A3"  
#>  [37] "COL8A2"   "COMP"     "COX7A1"   "CSF1R"    "CXCL12"   "CXCL14"  
#>  [43] "CXCL9"    "DCN"      "DDR2"     "DIO2"     "ECM2"     "EDIL3"   
#>  [49] "EDNRA"    "EGFL6"    "EMCN"     "ENPEP"    "ENPP2"    "ERG"     
#>  [55] "F13A1"    "FAP"      "FASLG"    "FBLN2"    "FCGR2A"   "FCGR2B"  
#>  [61] "FMO1"     "FOXF1"    "FPR1"     "FRZB"     "GIMAP5"   "GREM1"   
#>  [67] "HDC"      "HEPH"     "HGF"      "IGF1"     "IGSF6"    "IL18R1"  
#>  [73] "IL1B"     "ISLR"     "ITGAM"    "ITGBL1"   "ITIH3"    "ITIH5"   
#>  [79] "ITM2A"    "KCNJ8"    "KDR"      "LDB2"     "LMOD1"    "LRRC15"  
#>  [85] "LRRC32"   "LUM"      "LY86"     "MAF"      "MFAP5"    "MMP3"    
#>  [91] "MS4A4A"   "MS4A6A"   "MSR1"     "MXRA5"    "NME8"     "NOX4"    
#>  [97] "OLFML1"   "OLFML2B"  "PAPPA"    "PCDH12"   "PDE2A"    "PDGFRB"  
#> [103] "PIK3R5"   "PLPPR4"   "PLXDC1"   "PLXNC1"   "PRKG1"    "PTGER3"  
#> [109] "PTGIS"    "RAMP3"    "RARRES2"  "RASGRP3"  "RGS4"     "RSAD2"   
#> [115] "RUNX1T1"  "SAMSN1"   "SCUBE2"   "SERPING1" "SFRP4"    "SGCD"    
#> [121] "SH2D1A"   "SIGLEC1"  "SP140"    "SPON1"    "SPON2"    "SULF1"   
#> [127] "TCF21"    "TENM4"    "TFEC"     "THBS2"    "TLR2"     "TLR7"    
#> [133] "TNFSF4"   "TNN"      "TRAT1"    "VCAM1"    "VSIG4"    "WNT2"    
#> [139] "ZEB2"     "ZFPM2"    "ZNF423"  
#>