Display available gene signatures

get_sig_info() returns information about all gene signatures implemented in hacksig.

Usage

get_sig_info()

Value

A tibble with one row per signature and four columns:

signature_id, a unique identifier associated to a signature;

signature_keywords, valid keywords to use in the signatures argument of hack_sig() and check_sig() as well as in the keywords argument of get_sig_genes();

publication_doi, the original publication DOI;

description, a brief description about the signature.

See also

Examples

get_sig_info()
#> # A tibble: 40 × 4
#>    signature_id        signature_keywords            publication_doi description
#>    <chr>               <chr>                         <chr>           <chr>      
#>  1 ayers2017_immexp    ayers2017_immexp|immune expa… 10.1172/JCI911… Immune exp…
#>  2 bai2019_immune      bai2019_immune|head and neck… 10.1155/2019/3… Immune/inf…
#>  3 cinsarc             cinsarc|metastasis|sarcoma|s… 10.1038/nm.2174 Biomarker …
#>  4 dececco2014_int172  dececco2014_int172|head and … 10.1093/annonc… Signature …
#>  5 eschrich2009_rsi    eschrich2009_rsi|radioresist… 10.1016/j.ijro… Genes aime…
#>  6 estimate_immune     estimate|immune               10.1038/ncomms… Immune-rel…
#>  7 estimate_stromal    estimate|stromal              10.1038/ncomms… Stroma-rel…
#>  8 eustace2013_hypoxia eustace2013_hypoxia|head and… 10.1158/1078-0… Hypoxia-re…
#>  9 fan2021_ferroptosis ferroptosis|head and neck sq… 10.3389/fgene.… A model of…
#> 10 fang2021_irgs       fang2021_irgs|immune          10.18632/aging… Immune-rel…
#> # ℹ 30 more rows