Hack the ESTIMATE scores

Obtain Immune, Stroma, ESTIMATE and Tumor Purity scores from a cohort of samples, using the method implemented in Yoshihara et al., 2013.

Usage

hack_estimate(expr_data)

Source

bioinformatics.mdanderson.org/public-software/estimate/

Arguments

expr_data

A normalized gene expression matrix (or data frame) with gene symbols as row names and samples as columns.

Value

A tibble with one row for each sample in expr_data and five columns: sample_id, immune_score, stroma_score, estimate_score and purity_score.

Details

The ESTIMATE (Estimation of STromal and Immune cells in MAlignant Tumors using Expression data) method was developed with the aim to estimate the fraction of tumor cells in a sample by using gene expression instead of copy number data. The fundamental assumption of this method is that the tumor microenvironment is a very rich and dynamic ecosystem, in which immune infiltrating cells and stroma play a major role. The ESTIMATE score is defined as the combination (i.e. sum) of immune and stroma scores and can be thought of as a "non-tumor score". Consequently, a high ESTIMATE enrichment gives a low tumor purity score and viceversa.

Algorithm

Raw immune and stromal signatures scores are computed using single sample GSEA with rank normalization (Barbie et al., 2009). Then, the ESTIMATE score is computed by summing the immune and stroma scores. Finally, the tumor purity score is obtained with the following formula: $$Purity = cos(0.6049872018 + 0.0001467884 * ESTIMATE)$$

References

Barbie, D. A., Tamayo, P., Boehm, J. S., Kim, S. Y., Moody, S. E., Dunn, I. F., Schinzel, A. C., Sandy, P., Meylan, E., Scholl, C., Fröhling, S., Chan, E. M., Sos, M. L., Michel, K., Mermel, C., Silver, S. J., Weir, B. A., Reiling, J. H., Sheng, Q., Gupta, P. B., … Hahn, W. C. (2009). Systematic RNA interference reveals that oncogenic KRAS-driven cancers require TBK1. Nature, 462(7269), 108–112. doi:10.1038/nature08460 .

Yoshihara, K., Shahmoradgoli, M., Martínez, E., Vegesna, R., Kim, H., Torres-Garcia, W., Treviño, V., Shen, H., Laird, P. W., Levine, D. A., Carter, S. L., Getz, G., Stemke-Hale, K., Mills, G. B., & Verhaak, R. G. (2013). Inferring tumour purity and stromal and immune cell admixture from expression data. Nature communications, 4, 2612. doi:10.1038/ncomms3612 .

Examples

hack_estimate(test_expr)
#> # A tibble: 20 × 5
#>    sample_id immune_score stroma_score estimate_score purity_score
#>    <chr>            <dbl>        <dbl>          <dbl>        <dbl>
#>  1 sample1          -636.         778.           142.        0.811
#>  2 sample10         1590.        1297.          2887.        0.516
#>  3 sample11         2040.         512.          2552.        0.557
#>  4 sample12         1835.         772.          2607.        0.551
#>  5 sample13          632.         778.          1409.        0.688
#>  6 sample14         1185.        1005.          2191.        0.601
#>  7 sample15         2393.         415.          2808.        0.526
#>  8 sample16         1308.        1274.          2582.        0.554
#>  9 sample17         1181.         677.          1858.        0.639
#> 10 sample18          851.        1517.          2368.        0.580
#> 11 sample19         1672.         980.          2652.        0.545
#> 12 sample2          2118.         703.          2821.        0.524
#> 13 sample20         1639.        2482.          4121.        0.353
#> 14 sample3           725.         805.          1530.        0.675
#> 15 sample4           737.        2031.          2768.        0.531
#> 16 sample5           181.        1129.          1310.        0.699
#> 17 sample6          1221.        1175.          2395.        0.576
#> 18 sample7          1322.         375.          1697.        0.657
#> 19 sample8           515.        1158.          1673.        0.660
#> 20 sample9           297.        1147.          1443.        0.685