Display available gene signatures

get_sig_info() returns information about all gene signatures implemented in hacksig.

Usage

get_sig_info()

Value

A tibble with one row per signature and four columns:

signature_id, a unique identifier associated to a signature;

signature_keywords, valid keywords to use in the signatures argument of hack_sig() and check_sig() as well as in the keywords argument of get_sig_genes();

publication_doi, the original publication DOI;

description, a brief description about the signature.

See also

Examples

get_sig_info()
#> # A tibble: 40 × 4
#>    signature_id        signature_keywords                        publi…¹ descr…²
#>    <chr>               <chr>                                     <chr>   <chr>  
#>  1 ayers2017_immexp    ayers2017_immexp|immune expanded          10.117… Immune…
#>  2 bai2019_immune      bai2019_immune|head and neck squamous ce… 10.115… Immune…
#>  3 cinsarc             cinsarc|metastasis|sarcoma|sts            10.103… Biomar…
#>  4 dececco2014_int172  dececco2014_int172|head and neck squamou… 10.109… Signat…
#>  5 eschrich2009_rsi    eschrich2009_rsi|radioresistance|radiose… 10.101… Genes …
#>  6 estimate_immune     estimate|immune                           10.103… Immune…
#>  7 estimate_stromal    estimate|stromal                          10.103… Stroma…
#>  8 eustace2013_hypoxia eustace2013_hypoxia|head and neck squamo… 10.115… Hypoxi…
#>  9 fan2021_ferroptosis ferroptosis|head and neck squamous cell … 10.338… A mode…
#> 10 fang2021_irgs       fang2021_irgs|immune                      10.186… Immune…
#> # … with 30 more rows, and abbreviated variable names ¹​publication_doi,
#> #   ²​description